Add time to read to posts and improve date formatting
