+++ astro.config.mjs
@@ -5,25 +5,12 @@ import mdx from "@astrojs/mdx";
 import react from "@astrojs/react";
 import tailwind from "@astrojs/tailwind";
 
-
-const tilDefaultLayout = () => {
-  // @ts-ignore
-  return function (_, file) {
-    // get keys of the file.data.astro.frontmatter
-    const frontMatterKeys = Object.keys(file.data.astro.frontmatter);
-    // If 'author' is present in frontmatter, use blog post layout
-    // Otherwise, use TIL Layout
-    const defaultLayout = frontMatterKeys.includes("author") ? "../../../layouts/BlogPostLayout.astro" : "../../layouts/TilsLayout.astro";
-    console.log(`Using default layout: '${defaultLayout}'`);
-    file.data.astro.frontmatter.layout =
-      file.data.astro.frontmatter.layout || defaultLayout;
-  };
-};
+import { getDefaultLayout, remarkReadingTime } from './src/utils/remarkPlugins.mjs'
 
 // https://astro.build/config
 export default defineConfig({
   integrations: [mdx(), react(), tailwind()],
   markdown: {
-    remarkPlugins: [tilDefaultLayout],
+    remarkPlugins: [getDefaultLayout, remarkReadingTime],
   }
 });
