+++ package.json
@@ -16,12 +16,13 @@
     "astro": "5.1.10",
     "framer-motion": "11.18.2",
     "htmlparser2": "10.0.0",
+    "mdast-util-to-string": "^4.0.0",
     "prismjs": "1.29.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-helmet": "6.1.0",
     "react-icons": "5.4.0",
-    "reading-time": "1.5.0",
+    "reading-time": "^1.5.0",
     "remark-external-links": "8.0.0",
     "replace-in-file": "8.3.0",
     "slugify": "1.6.6",
