+++ src/layouts/TilsLayout.astro
@@ -4,15 +4,16 @@ import type { MarkdownLayoutProps } from 'astro';
 import { PostMetaData } from '../components/post-metadata'
 import React from 'react'
 import { Sharing } from '../components/sharing'
-import { makeTILUrl } from '../utils/routes'
 import { Pagination } from '../components/pagination'
+import { formatLongDate } from '../utils/dates'
 
 type Props = MarkdownLayoutProps<{
-  date: Date;
+  date: string;
   title: string;
 }>;
 const { frontmatter, url, file } = Astro.props;
 
+const postDate = new Date(frontmatter.date);
 const editUrl = `https://github.com/browniebroke/browniebroke.com/blob/master/src/${
     file.split("/src/")[1]
   }`;
@@ -20,7 +21,7 @@ const postPath = url || "";
 ---
 <BaseLayout>
   <h1 class="text-3xl font-bold mt-8 mb-4">{ frontmatter.title }</h1>
-  <PostMetaData dateTimeToRead={frontmatter.date} editUrl={editUrl} />
+  <PostMetaData dateTimeToRead={formatLongDate(postDate)} editUrl={editUrl} />
   <slot />
 
   <Sharing frontmatter={frontmatter} path={postPath} />
