+++ src/pages/blog.astro
@@ -15,7 +15,6 @@ const allBlogPostsByMostRecent = allBlogPosts
   .toSorted((p1, p2) => (p2.data.date - p1.data.date))
   .map(
     (post) => {
-      console.log(post);
       return ({
         title: post.data.title,
         description: post.data.description,
